Move helper functions out into own files
